Implemented autoupdates to Chromedriver and Geckodriver! woohoo!!
