+++ .gitignore
@@ -5,4 +5,4 @@ log.html
 output.xml
 report.html
 .gitignore
-libraries/__pycache__/webDriver.cpython-310.pyc
+Libraries/__pycache__/webDriver.cpython-310.pyc
