+++ PageObjects/Locators.py
@@ -1,0 +1,4 @@
+# Example Page Object Model
+txt_loginUserName = "name:userName"
+txt_loginPassword = "name:password"
+btn_signIn = "xpath://input[@name='login']"
